Extensible device API. Redesign with object-oriented style, making it easier to add other protocols
